Support writing OCSP to Redis on revocation
If a Redis client is configured for the SA service, OCSP responses created
during a revocation event will be written to Redis on a best effort basis.

Redis supports using an existing TTL stored with the Redis key. Add
support in rocsp so if the TTL duration passed to StoreResponse is -1h, it
sets `redis.KeepTTL` as the TTL value. We can make a high percentage
assumption that, upon revocation, the key storing the OCSP response for
that serial will exist in Redis and the TTL will be able to be reused. In
the case that it is creating the key for the first time, the revoked
response will be stored with no expiration, but the OCSP updater will
write a corrected TTL on the next update.

Fixes #5888