rocsp: add "connection_pool" to metric names
Most of our existing Redis metrics are relatd to the connection pool,
but some sound like they could be request-oriented (like redis_hits and
redis_misses). Fix that with better naming.

Also move the stat creation into rocsp/metrics.go so the names are in
the same file as the collector.