rocsp: add "connection_pool" to metric names (#6112)
Most of our existing Redis metrics are related to the connection pool,
but some sound like they could be request-oriented (like redis_hits and
redis_misses). Fix that with better naming. Also, join three related
metrics (hits, misses, and timeouts) under a single metric name with
labels so they more easily by grouped and counted.

Move the stat creation into rocsp/metrics.go so the names are in the
same file as the collector.

Add a unittest for rocsp metrics.