timeout connection failure: provide more details
"likely firewall problem" is a good description of the problem, but
since it's such an unusual situation, providing more info can ease
debugging.

Personal experience: if I knew this information, it would immediately
help me narrow down the problem to my router/ISP and not waste time with
IPv6 and DNS that were unrelated.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>