redis_source: improve metrics
This gives less confusing naming to the metrics in checked_redis_source
("revocation_re_sign_success" instead of "sign_and_save_success").

It also uses a new type as the `cause` parameter to signAndSave, to make
it clear what should be passed.

In redis_source, it splits `counter` into two separate Prometheus
counters: one for requests in general, and a separate one for
signAndSave. The counter for signAndSave has two labels: cause and
result.