Remove crlNumber from CRL S3 paths (#6352)
Change the S3 object paths to not include the CRL Number.

At one point, the plan was to upload all of the CRL shards to
S3 paths containing their CRL Number (which monotonically
increases every generation), and then later move or symlink
them into paths not containing that number. However, we saw
that S3 does not have any atomic move or rename semantics,
so we decided to instead enable object versioning and upload
the shards to the same path every time. Apparently I never fixed
the object key computation to match the updated design.

The CRL Number is still stored on the object as a metadata tag.