Stabilize CRL shard boundaries
Add two new config keys to the crl-updater:
* shardWidth, which controls the width of the chunks that we divide all
  of time into, with a default value of "16h" (approximately the same as
  today's shard width derived from 128 shards covering 90 days); and
* lookbackPeriod, which controls the amount of already-expired
  certificates that should be included in our CRLs to ensure that even
  certificates which are revoked immediately before they expire still
  show up in aborts least one CRL, with a default value of "24h"
  (approximately the same as today's lookback period derived from our
  run frequency of 6h).

Use these two new values to change the way CRL shards are computed.

Previously, we would compute the total time we care about based on the
configured certificate lifetime (to determine how far forward to look)
and the configured update period (to determine how far back to look),
and then divide that time evenly by the number of shards. However, this
method had two fatal flaws. First, if the certificate lifetime is
incorrect, then the CRL updater will fail to query the database for some
certs that should be included in the CRLs. Second, if the update period
is changed, this would change the lookback period, which in turn would
change the shard width, causing all CRL entries to suddenly change which
shard they're in.

Instead, first compute all chunk locations based only on the shard width
and number of shards. Then determine which chunks we need to care about
based on the configured lookback period and by querying the database for
the farthest-future expiration, to ensure we cover all extant
certificates. This may mean that more than one chunk of time will get
mapped to a single shard, but that's okay -- each chunk will remain
mapped to the same shard for the whole time we care about it.

Fixes #6438
Fixes #6440