SA.NewOrder: Get order status inside transaction
The SA's NewOrder and NewOrderAndAuthzs methods both write new rows
(the order itself, new authorizations, and new OrderToAuthz relations)
to the database, and then quickly turn around and query the database for
a bunch of authz rows to compute the status of the new order. This is
necessary because many orders are created already referencing existing
authorizations, and the state of those authorizations is not known at
order creation time, but does affect the order's status.

Due to replication lag, it can cause issues if the database writes go
to the primary database, but the follow-up read goes to a read-only
database replica which may be lagging slightly. To prevent this issue,
conduct the reads on the same transaction as the writes.